+++ CHANGELOG.md
@@ -20,6 +20,25 @@ Types of changes
 - `Security` in case of vulnerabilities.
 
 -->
+## [0.4.167](#409)
+
+### Changed
+- Docker executor contianer run pulls image if it doesn't exist locally
+- Readme update on testing
+
+### Added
+- Executor tests and test data fetch using make
+
+## [0.4.166](#396)
+
+### Added
+- Firecracker executor
+
+## [0.4.165](#346)
+
+### Added
+- Generic executor package interface and its docker implementation
+
 ## [0.4.164](#340)
 
 ### Fixed